requirements.txt
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
Flask==3.0.3
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
linkedin_api==2.1.1
lxml==5.2.2
MarkupSafe==2.1.5
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
oauthlib==3.2.2
packaging==24.0
portalocker==2.8.2
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
pywin32==306
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
